requirements.txt
altair
pandas
streamlit

numpy
pyarrow
functions_framework
langchain==0.0.139
fuzzywuzzy==0.18.0
openai==0.27.0
tqdm==4.64.1

chromadb==0.3.21
tiktoken==0.3.3
faiss-cpu==1.7.3
# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

# GCP
google-cloud-bigquery-storage==2.13.0
google-cloud-bigquery==2.34.2
google-cloud-storage

